+++ src/Matrix.java
@@ -1,16 +1,16 @@
 public class Matrix {
     private Integer rows;
     private Integer columns;
-    private Integer[][] matrix;
+    private Double[][] matrix;
 
     public Matrix(int rows, int columns) {
         this.rows = rows;
         this.columns = columns;
-        this.matrix = new Integer[rows][columns];
+        this.matrix = new Double[rows][columns];
 
         for (int i = 0; i < rows; i++) {
             for (int j = 0; j < columns; j++) {
-                matrix[i][j] = 0;
+                matrix[i][j] = 0.0;
             }
         }
     }
@@ -23,14 +23,14 @@ public class Matrix {
         return columns;
     }
 
-    public void setValue(int row, int column, int value) {
+    public void setValue(int row, int column, double value) {
         if (!isValidIndex(row, column)) {
             throw new IllegalArgumentException("Invalid position!");
         }
         matrix[row][column] = value;
     }
 
-    public Integer getValue(int row, int column) {
+    public double getValue(int row, int column) {
         if (isValidIndex(row, column)) {
             return matrix[row][column];
         } else {
